# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: CampusLive API
  version: 1.0.0
  description: |
    Realtime "airport board" for a university building. Status/phase logic lives in the Go engine;
    clients only derive progress and countdowns from the timestamps in a Snapshot.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/buildings/{code}/map:
    get:
      tags: [buildings]
      operationId: getBuildingMap
      summary: Floors and rooms with geometry (cached, ETag)
      parameters:
        - $ref: '#/components/parameters/BuildingCode'
      responses:
        '200':
          description: Map
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BuildingMap' }
        '304': { description: Not modified }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/buildings/{code}/board:
    get:
      tags: [schedule]
      operationId: getBoard
      summary: Snapshot at a moment (server now by default) — used for SSR and time-travel
      parameters:
        - $ref: '#/components/parameters/BuildingCode'
        - name: at
          in: query
          description: Instant to compute the snapshot for. Defaults to the server clock.
          schema: { type: string, format: date-time }
        - name: date
          in: query
          description: Local date (YYYY-MM-DD) of the timeline to materialise. Defaults to the local date of `at`.
          schema: { type: string, format: date }
      responses:
        '200':
          description: Snapshot
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snapshot' }
        '304': { description: Not modified }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/buildings/{code}/timeline:
    get:
      tags: [schedule]
      operationId: getTimeline
      summary: All sessions of a day (time-travel bar, room panel)
      parameters:
        - $ref: '#/components/parameters/BuildingCode'
        - $ref: '#/components/parameters/Date'
      responses:
        '200':
          description: Day timeline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DayTimeline' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/rooms/{code}/day:
    get:
      tags: [schedule]
      operationId: getRoomDay
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Date'
      responses:
        '200':
          description: Sessions of the room for the day
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DaySchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/teachers/{id}/day:
    get:
      tags: [schedule]
      operationId: getTeacherDay
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/Date'
      responses:
        '200':
          description: Sessions of the teacher for the day
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DaySchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/groups/{code}/day:
    get:
      tags: [schedule]
      operationId: getGroupDay
      parameters:
        - name: code
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Date'
      responses:
        '200':
          description: Sessions of the group for the day
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DaySchedule' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/events:
    get:
      tags: [realtime]
      operationId: events
      summary: |
        Server-Sent Events stream. Events: `snapshot` (Snapshot), `announcement` (Announcement), `heartbeat` ({at}).
        A full `snapshot` is sent on connect and after every transition or override.
      parameters:
        - name: building
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema: { type: string }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/overrides:
    post:
      tags: [admin]
      operationId: createOverride
      security: [{ ApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OverrideRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Override' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/overrides/{id}:
    delete:
      tags: [admin]
      operationId: deleteOverride
      security: [{ ApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/announcements:
    post:
      tags: [admin]
      operationId: createAnnouncement
      security: [{ ApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnnouncementRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Announcement' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/admin/catalog:
    get:
      tags: [admin]
      operationId: getAdminCatalog
      summary: Reference data for the schedule editor (semesters, slots, schedulable rooms, teachers, courses, groups)
      security: [{ ApiKey: [] }]
      parameters:
        - name: building
          in: query
          schema: { type: string, default: A }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCatalog' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/admin/lessons:
    get:
      tags: [admin]
      operationId: listLessons
      summary: Lesson templates (the weekly timetable) of a semester, optionally of one room
      security: [{ ApiKey: [] }]
      parameters:
        - name: building
          in: query
          schema: { type: string, default: A }
        - name: semesterId
          in: query
          description: Defaults to the semester containing today
          schema: { type: string, format: uuid }
        - name: roomCode
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Lesson' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [admin]
      operationId: createLesson
      summary: Add a weekly lesson; the board of every open screen updates over SSE
      security: [{ ApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LessonRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lesson' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: The room, teacher or a group is already busy in that slot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/v1/admin/lessons/{id}:
    delete:
      tags: [admin]
      operationId: deleteLesson
      security: [{ ApiKey: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    BuildingCode:
      name: code
      in: path
      required: true
      schema: { type: string }
    Date:
      name: date
      in: query
      description: Local date (YYYY-MM-DD). Defaults to today in the building timezone.
      schema: { type: string, format: date }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    BuildingMap:
      type: object
      required: [building, name, timezone, viewBox, floors]
      properties:
        building: { type: string }
        name: { type: string }
        timezone: { type: string }
        viewBox:
          type: array
          items: { type: number }
          minItems: 4
          maxItems: 4
        floors:
          type: array
          items: { $ref: '#/components/schemas/MapFloor' }
...
    Snapshot:
      type: object
      required: [building, at, date, nextTransitionAt, stats, rooms, now, next, announcements]
      properties:
        building: { type: string }
        at: { type: string, format: date-time }
        date: { type: string, format: date, description: Local date the timeline was materialised for }
        nextTransitionAt: { type: string, format: date-time, nullable: true }
        stats: { $ref: '#/components/schemas/SnapshotStats' }
        rooms:
          type: array
          items: { $ref: '#/components/schemas/RoomLiveState' }
        now:
          type: array
          description: Sessions in phase live|ending, sorted by endAt
          items: { $ref: '#/components/schemas/SessionView' }
        next:
          type: array
          description: Sessions in phase soon|upcoming starting within the NEXT horizon (+ cancelled/moved in the window), sorted by startAt
          items: { $ref: '#/components/schemas/SessionView' }
        announcements:
          type: array
          items: { $ref: '#/components/schemas/Announcement' }
    DayTimeline:
      type: object
      required: [building, date, at, slots, sessions]
      properties:
        building: { type: string }
        date: { type: string, format: date }
        at: { type: string, format: date-time }
        slots:
          type: array
          items: { $ref: '#/components/schemas/TimeSlot' }
        sessions:
          type: array
          items: { $ref: '#/components/schemas/SessionView' }
...
    DaySchedule:
      type: object
      required: [date, at, subject, sessions]
      properties:
        date: { type: string, format: date }
        at: { type: string, format: date-time }
        subject:
          type: object
          required: [kind, id, label]
          properties:
            kind: { type: string, enum: [room, teacher, group] }
            id: { type: string }
            label: { type: string }
        sessions:
          type: array
          items: { $ref: '#/components/schemas/SessionView' }
...
    OverrideRequest:
      type: object
      required: [kind, date]
      properties:
        kind: { $ref: '#/components/schemas/OverrideKind' }
        date: { type: string, format: date }
        lessonId: { type: string, format: uuid, nullable: true, description: Required for every kind except extra }
        newRoomCode: { type: string, nullable: true, description: move }
        newTeacherId: { type: string, format: uuid, nullable: true, description: reassign_teacher }
        delayMinutes: { type: integer, nullable: true, minimum: 1, maximum: 180, description: delay }
        courseCode: { type: string, nullable: true, description: extra }
        roomCode: { type: string, nullable: true, description: extra }
        teacherId: { type: string, format: uuid, nullable: true, description: extra }
        slotIdx: { type: integer, nullable: true, description: extra }
        groups:
          type: array
          nullable: true
          items: { type: string }
          description: extra
        note: { type: string, nullable: true }
    Override:
      type: object
      required: [id, kind, date, lessonId, createdAt]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/OverrideKind' }
        date: { type: string, format: date }
        lessonId: { type: string, format: uuid, nullable: true }
        newRoomCode: { type: string, nullable: true }
        newTeacherId: { type: string, format: uuid, nullable: true }
        delayMinutes: { type: integer, nullable: true }
        note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...